bump iced from 0.2.0 to 0.3.0

+++ src/gui/scrollables.rs
@@ -18,7 +18,7 @@ impl Variant {
                 // TODO: fix this when proper patch is added to iced.
                 // See: https://github.com/hecrj/iced/pull/607#issuecomment-742818666
                 s = unsafe {
-                    let mut tmp = std::mem::transmute::<_, (Option<f32>, f32)>(s);
+                    let mut tmp = std::mem::transmute::<_, (Option<f64>, f64)>(s);
                     tmp.1 = 999999.0;
                     std::mem::transmute::<_, scrollable::State>(tmp)
                 };
